fix(dashboard): harden clipboard copy for non-secure contexts (#244)
Templates.tsx copyName used raw navigator.clipboard.writeText, which is undefined
over plain HTTP on a LAN IP (secure-context-only) — the copy button would throw an
unhandled rejection and silently fail, reintroducing the bug fixed in #244.

Extract the guarded copy (navigator.clipboard with an execCommand('copy') fallback)
into a shared utils/clipboard.ts and use it from both Templates.tsx and ApiKeys.tsx
(which previously carried its own inline copy), removing the duplication.
